Rename base-exercises folder
