+++ src/app/exercise/exercises/utility/base-exercises/BaseCommonSettingsExercise.ts
