+++ src/app/exercise/exercises/utility/base-exercises/BaseExercise.ts
