+++ src/app/exercise/exercises/utility/base-exercises/BaseTonalChordProgressionExercise.ts
