+++ src/app/exercise/exercises/utility/base-exercises/BaseTonalExercise.ts
